feat: Add version match checking to handshake

+++ game/multiplayer/multiplayer.h
@@ -6,7 +6,7 @@ void pc_multi_disconnect();
 
 void pc_multi_setup_host();
 void pc_multi_setup_client(u32 ip_ptr, u32 port);
-int pc_multi_get_status();
+int64_t pc_multi_get_status();
 void pc_multi_request_full_sync();
 void pc_multi_stop_search();
 void pc_multi_start_search();
